Refactored Renting Logic
Refactored the renting logic by separating out parts of it into a "Rentable" interface. This enables easy renting of apartments or properties as a whole. In tow changes to naming schemes were applied to clearly separate the concepts of living spaces (ApartmentUnits and Properties (which in turn could be Houses, Offices or Factories - only the tenant will care about this difference) and properties containing living spaces (ApartmentComplex).