Refactored Application Core Concept
The domain layers aggregate root was changed from the existing RentalAgreement version to an updated implementations of the Rentable class. ApartmentUnit and Property both got upgraded to aggregates from entities, meanwhile RentalAgreement was demoted.
This change was done to fix a short sighted issue that would have led to data loss of currently not rented properties upon de- and reserialization. The application core now aligns to the real life domain context more closely.